Extract `RouteDecisionHandler` in a separate file.
